pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-mario-the-plumber"
version = "0.1.0"
description = "OpenEnv environment for repairing broken ETL pipelines"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.0",
]

[project.scripts]
server = "server.app:main"

[tool.uv]
package = false

[tool.setuptools]
include-package-data = true
py-modules = ["models", "client", "inference"]

[tool.setuptools.packages.find]
include = ["server*", "benchmark*", "scripts*"]
